Add unit test for CloseTimerTask,ReconnectTimerTask,HeaderExchangeServer (#9046)

+++ dubbo-remoting/dubbo-remoting-api/src/main/java/org/apache/dubbo/remoting/exchange/support/header/HeaderExchangeServer.java
@@ -114,7 +114,7 @@ public class HeaderExchangeServer implements ExchangeServer {
         }
         startClose();
         if (timeout > 0) {
-            final long max = (long) timeout;
+            final long max = timeout;
             final long start = System.currentTimeMillis();
             if (getUrl().getParameter(Constants.CHANNEL_SEND_READONLYEVENT_KEY, true)) {
                 sendChannelReadOnlyEvent();
